requirements.txt
torch
tqdm
matplotlib
numpy
torchvision
pillow
